supportTask/cal_aqi/requirements.txt
numpy==1.18.1
pandas==1.0.1
python-aqi==0.6.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
